+++ src/app/data/poe-ninja/service/item-overview-http.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { BrowserService, LoggerService, SessionService } from '@app/service';
+import { BrowserService, LoggerService } from '@app/service';
 import { environment } from '@env/environment';
 import { Observable, of, throwError } from 'rxjs';
 import { delay, flatMap, retryWhen } from 'rxjs/operators';
@@ -56,7 +56,6 @@ export class ItemOverviewHttpService {
     constructor(
         private readonly httpClient: HttpClient,
         private readonly browser: BrowserService,
-        private readonly session: SessionService,
         private readonly logger: LoggerService) {
         this.baseUrl = `${environment.poeNinja.baseUrl}/api/data/itemoverview`;
     }
@@ -95,7 +94,7 @@ export class ItemOverviewHttpService {
             case 403:
                 return this.browser.retrieve(url).pipe(delay(RETRY_DELAY));
             default:
-                return this.session.clear().pipe(delay(RETRY_DELAY));
+                return of(null).pipe(delay(RETRY_DELAY));
         }
     }