+++ src/app/shared/module/poe/service/index.ts
@@ -6,6 +6,7 @@ export * from './item/item-price-prediction.service';
 export * from './item/item-search-analyze.service';
 export * from './item/item-search.service';
 export * from './item/item.service';
+export * from './item/processor/item-processor.service';
 export * from './leagues.service';
 export * from './stash/stash.service';
 export * from './stats/stats.service';